Unify _to_date converter and add git workflow guidelines
- Extended base _to_date function to handle Union[str, datetime, date]
- Removed duplicate _to_date implementations from arrow and polars converters
- Added git workflow guidelines to CLAUDE.md (never commit directly to master)
